requirements.txt
appdirs==1.4.3
astroid==2.4.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
docker==4.2.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
lockfile==0.12.2
lorem==0.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pika==1.1.0
progress==1.5
pylint==2.5.3
pyparsing==2.4.6
python-dotenv==0.13.0
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
websocket-client==0.57.0
wrapt==1.12.1
